Added Healthbar, and Spikes
I also tweaked the strength of the rod.
Spikes still need iframes, and there needs to be a way to see the current health number rather than percentage.
Death is really simple right now, but a full scene reset is probably what we should do in the future to fix issues of hooks and other things persisting.