Added hook model and placeholder line renderer.
The line render is a bit buggy due to physics interpolation. We'll need
to find a better solution later. I think a billboarded and rotated quad
could work well since the renderer should deal with the projection
better than just manually unprojecting the positions every frame (this
is what is causing the jitteryness).